common.c (COM_InitArgv): remove the trailing whitespace added at the end of cmdline string only if there is one. print out the command line as an informational message.
git-svn-id: svn://svn.code.sf.net/p/quakespasm/code/trunk/quakespasm@19 af15c1b1-3010-417e-b628-4374ebc0bcbd

+++ Quake/common.c
@@ -1143,7 +1143,10 @@ void COM_InitArgv (int argc, char **argv)
 			break;
 	}
 
-	com_cmdline[n-1] = 0; //johnfitz -- kill the trailing space
+	if (com_cmdline[n-1] = ' ')
+		com_cmdline[n-1] = 0; //johnfitz -- kill the trailing space
+
+	Con_Printf("\nCommand line: %s",com_cmdline);
 
 	safe = false;
